don't use System.stacktrace()
Accessing the stacktrace outside of a rescue/catch is
deprecated. Adding a deprecation as of now. Also erlang:get_stacktrace
is deprecated (which is what elixir is calling now) and from OTP 23
onwards get_stacktrace will return `[]` and it will be removed in OTP
24.

The get_stacktrace call used in logger is a stop gap solution. Proper
solution might need major changes to logger module